% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_cf_mgn.R
\name{cfsens_cf_mgn}
\alias{cfsens_cf_mgn}
\title{Robust conformal counterfactual inference with marginal guarantee}
\usage{
cfsens_cf_mgn(
  X,
  Y,
  T,
  Gamma,
  alpha,
  side = c("two", "above", "below"),
  score_type = c("cqr"),
  ps_fun = regression_forest,
  ps = NULL,
  pred_fun = quantile_forest,
  train_prop = 0.75,
  train_id = NULL
)
}
\arguments{
\item{X}{covariates.}

\item{Y}{the observed outcome vector.}

\item{T}{the vector of treatment assignments.}

\item{Gamma}{The confounding level.}

\item{alpha}{the target confidence level.}

\item{side}{the type of predictive intervals that takes value in \{"two", "above", "below"\}. See details.}

\item{score_type}{the type of nonconformity scores. The default is "cqr".}

\item{ps_fun}{a function that models the treatment assignment mechanism. The default is "regression_forest".}

\item{ps}{a vector of propensity score. The default is \code{NULL}.}

\item{pred_fun}{a function that models the potential outcome conditional on the covariates. The default is "quantile_forest".}

\item{train_prop}{proportion of units used for training. The default is 75\\%.}

\item{train_id}{The index of the units used for training. The default is \code{NULL}.}
}
\value{
an \code{cfmgn} object.
}
\description{
\code{cfsens_cf_mgn} constructs robust predictive intervals for
counterfactuals with the margianl guarantee.
}
\details{
When \code{side = "two"}, the predictive interval takes the form \link{a,b};
when \code{side = "above"}, the predictive interval takes the form [a,Inf);
when \code{side = "below"}, the predictive interval takes the form (-Inf,a].

When \code{null_type = "sharp"}, the null hypothesis is H_0: Y(1) - Y(0) = 0.
When \code{null_type = "negative"}, the null hypothesis is H_0: Y(1) - Y(0) <= 0.
When \code{null_type = "positive"}, the null hypothesis is H_0: Y(1) - Y(0) >= 0.

[a,Inf);
when 47kTFQ8ra99olZKWMot6xGQUo84wvvrn-3-, the predictive interval takes the form (-Inf,a]: R:a,Inf);\%0Awhen\%2047kTFQ8ra99olZKWMot6xGQUo84wvvrn-3-,\%20the\%20predictive\%20interval\%20takes\%20the\%20form\%20(-Inf,a
}
\examples{



}
\seealso{
\code{\link{cfsens_cf_pac}}
}
