% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_kstar}
\alias{find_kstar}
\title{Find the cutoff k* in the marginal conformal precedure}
\usage{
find_kstar(score, lx, ux, ux_test, alpha)
}
\description{
Find the cutoff k* in the marginal conformal precedure
}
