% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lr_bnds}
\alias{lr_bnds}
\title{Compute the bounds for the likelihood ratio under the marginal sensitivity model}
\usage{
lr_bnds(
  estimand = c("treated", "control"),
  type = c("ate", "att", "atc"),
  Gamma,
  ps,
  p1,
  p0
)
}
\description{
Compute the bounds for the likelihood ratio under the marginal sensitivity model
}
