% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_cf_mgn.R
\name{predict.cfmgn}
\alias{predict.cfmgn}
\title{Prediction method for cfmgn objects}
\usage{
\method{predict}{cfmgn}(
  obj,
  X_test,
  estimand = c("treated", "control"),
  type = c("ate", "att", "atc")
)
}
\arguments{
\item{obj}{an object of class \code{cfmgn}.}

\item{X_test}{testing covariates.}

\item{estimand}{the inferential target that
takes value in \{"treated", "control"\}. See details.}

\item{type}{the type of inference target.
Takes value in \{"ate", "att", "atc"\}. See details.}
}
\value{
predictive intervals. A data.frame that contains \code{nrow(X_test)} rows and
two columns: "Y_hi" refers the upper bound and "Y_lo"
the lower bound.
}
\description{
Obtains predictive intervals on a test dataset based on
an \code{cfmgn} object from \code{link{cfsens_cf_mgn}}.
}
\details{
When \code{type = "treated"}, predictive intervals for Y(1)
are constructed; when \code{type = "control"}, predictive intervals
for Y(0) are constructed.

When \code{type = "ate"}, the inference is valid unconditionally;
when \code{type = "att"}, the inference is valid conditional on T=1, and
\code{Y1_test} should be provided;
when \code{type = "atc"}, the inference is valid conditional on T=0, and
\code{Y0_test} should be provided.
}
