% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_cf_pac.R
\name{predict.cfpac}
\alias{predict.cfpac}
\title{Predictive interval for cfpac objects}
\usage{
\method{predict}{cfpac}(
  obj,
  X_test,
  estimand = c("treated", "control"),
  type = c("ate", "att", "atc"),
  bnd_type = c("wsr", "hoeffding"),
  num_grids = 1000,
  rand_ind = NULL,
  wsr_seed = 24601
)
}
\arguments{
\item{obj}{an object of class \code{cfpac}.}

\item{X_test}{testing covariates.}

\item{estimand}{the inferential target that
takes value in \{"treated", "control"\}. See details.}

\item{type}{the type of inference target.
Takes value in \{"ate", "att", "atc"\}. See details.}
}
\value{
predictive intervals. A data.frame that contains \code{nrow(X_test)} rows and
two columns: "Y_hi" refers the upper bound and "Y_lo"
the lower bound.
}
\description{
Obtains predictive intervals on a test dataset based on
an \code{cfpac} object from \code{link{cfsens_cf_pac}}.
}
\details{
When \code{type = "treated"}, predictive intervals for Y(1)
are constructed; when \code{type = "control"}, predictive intervals
for Y(0) are constructed.

When \code{type = "ate"}, the inference is valid unconditionally;
when \code{type = "att"}, the inference is valid conditional on T=1, and
\code{Y1_test} should be provided;
when \code{type = "atc"}, the inference is valid conditional on T=0, and
\code{Y0_test} should be provided.
}
