% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_mgn.R
\name{predict.itemgn}
\alias{predict.itemgn}
\title{Prediction method for itemgn objects}
\usage{
\method{predict}{itemgn}(
  obj,
  X_test,
  Y1_test = NULL,
  Y0_test = NULL,
  type = c("ate", "att", "atc"),
  Gamma_max = 5,
  gamma_length = 51
)
}
\arguments{
\item{obj}{an object of class \code{itemgn}.}

\item{X_test}{testing covariates.}

\item{Y1_test}{the potential outcome when treated.
The default is \code{NULL}. See details.}

\item{Y0_test}{the potential outcome when not treated.
The default is \code{NULL}. See details.}

\item{type}{the type of inference target. Takes value in \{"ate", "att", "atc"\}. See details.}

\item{Gamma_max}{the maximum value of Gamma to be considered for sensitivity analysis. The default is 5.}

\item{gamma_length}{the number of Gamma to be considered for sensitivity analysis. The default is 51.}
}
\value{
a vector of gamma-values.
}
\description{
Obtains gamma-values on a test dataset based on
an \code{itemgn} object from \code{link{cfsens_mgn}}.
}
\details{
When \code{type = "ate"}, the inference is valid unconditionally;
when \code{type = "att"}, the inference is valid conditional on T=1, and
\code{Y1_test} should be provided;
when \code{type = "atc"}, the inference is valid conditional on T=0, and
\code{Y0_test} should be provided.
}
