% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_pac.R
\name{predict.itepac}
\alias{predict.itepac}
\title{Prediction method for itepac objects}
\usage{
\method{predict}{itepac}(
  obj,
  X_test,
  Y1_test = NULL,
  Y0_test = NULL,
  type = c("ate", "att", "atc"),
  bnd_type = c("wsr"),
  num_grids = 1000,
  rand_ind = NULL,
  wsr_seed = 24601,
  max_gamma = 5,
  gamma_length = 51
)
}
\arguments{
\item{obj}{an object of class \code{itepac}.}

\item{X_test}{testing covariates.}

\item{Y1_test}{the potential outcome when treated.
The default is \code{NULL}. See details.}

\item{Y0_test}{the potential outcome when not treated.
The default is \code{NULL}. See details.}

\item{type}{the type of inference target. Takes value in \{"ate", "att", "atc"\}. See details.}

\item{bnd_type}{the method for constructing lower bound for the CDFs.
The default is "wsr".}

\item{gamma_length}{the number of Gamma to be considered for sensitivity analysis. The default is 51.}

\item{Gamma_max}{the maximum value of Gamma to be considered for sensitivity analysis. The default is 5.}
}
\value{
a vector of gamma-values.
}
\description{
Obtains gamma-values on a test dataset based on
an \code{itepac} object from \code{link{cfsens_pac}}.
}
\details{
When \code{type = "ate"}, the inference is valid unconditionally;
when \code{type = "att"}, the inference is valid conditional on T=1, and
\code{Y1_test} should be provided;
when \code{type = "atc"}, the inference is valid conditional on T=0, and
\code{Y0_test} should be provided.
}
