% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wsr_lower}
\alias{wsr_lower}
\title{Computing the lower bounds for G(t) via the Waudby-Smith and Ramdas inequality}
\usage{
wsr_lower(delta, x, num_grids)
}
\description{
Computing the lower bounds for G(t) via the Waudby-Smith and Ramdas inequality
}
