% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wsr_qtl}
\alias{wsr_qtl}
\title{Bi-search the cutoff for nonconformity scores}
\usage{
wsr_qtl(data_calib, delta, alpha, M, num_grids, rand_ind)
}
\description{
Bi-search the cutoff for nonconformity scores
}
