% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsens_pac.R
\name{cfsens_pac}
\alias{cfsens_pac}
\title{Sensitivity analysis of individual treatment effects with PAC-type guarantees}
\usage{
cfsens_pac(
  X,
  Y,
  T,
  alpha,
  delta,
  null_type = c("sharp", "negative", "positive"),
  score_type = c("cqr"),
  ps_fun = regression_forest,
  ps = NULL,
  pred_fun = quantile_forest,
  train_prop = 0.75,
  train_id = NULL
)
}
\arguments{
\item{X}{covariates.}

\item{Y}{the observed outcome vector.}

\item{T}{the vector of treatment assignments.}

\item{alpha}{the target confidence level.}

\item{delta}{the confidence level over the randomness over the calibration set.}

\item{null_type}{the null to be tested that takes value in \{"sharp", "negative", "positive"\}. See Details.}

\item{score_type}{the type of nonconformity scores. The default is "cqr".}

\item{ps_fun}{a function that models the treatment assignment mechanism. The default is "regression_forest".}

\item{ps}{a vector of propensity score. The default is \code{NULL}.}

\item{pred_fun}{a function that models the potential outcome conditional on the covariates. The default is "quantile_forest".}

\item{train_prop}{proportion of units used for training. The default is 75\\%.}

\item{train_id}{The index of the units used for training. The default is \code{NULL}.}
}
\value{
an \code{itepac} object.
}
\description{
\code{cfsens_pac} conducts sensitivity analysis of individual treatment effects
with PAC-type guarantee.
It currently supports testing the sharp null and the directional nulls.
}
\details{
When \code{null_type = "sharp"}, the null hypothesis is H_0: Y(1) - Y(0) = 0.
When \code{null_type = "negative"}, the null hypothesis is H_0: Y(1) - Y(0) <= 0.
When \code{null_type = "positive"}, the null hypothesis is H_0: Y(1) - Y(0) >= 0.
}
\examples{


}
\seealso{
\code{\link{cfsens_mgn}}
}
